+++ playground/app.vue
@@ -1,13 +1,12 @@
 <template>
   <UContainer class="min-h-screen flex items-center">
-    <UCard class="flex-1" :ui="{ background: 'bg-gray-50 dark:bg-gray-800/50', ring: 'ring-1 ring-gray-300 dark:ring-gray-700', divide: 'divide-y divide-gray-300 dark:divide-gray-700', header: { base: 'font-bold' } }">
+    <UCard class="flex-1" :ui="{ background: 'bg-transparent', ring: 'ring-1 ring-gray-300 dark:ring-gray-700', divide: 'divide-y divide-gray-300 dark:divide-gray-700', header: { base: 'font-bold' } }">
       <template #header>
         Welcome to the playground!
       </template>
 
-      <p class="text-gray-500 dark:text-gray-400">
-        Try your components here!
-      </p>
+      <UInput />
+      <UButton label="Log in" block class="mt-4" />
     </UCard>
   </UContainer>
 </template>