+++ scripts/git.sh
@@ -1,1 +1,1 @@
-git add . && git commit -m "`date`" && git push -u origin centlax
+git add . && git commit -m "buttons/inputs color change light-600 dark-500" && git push -u origin centlax