+++ src/runtime/ui.config/forms/input.ts
@@ -1,19 +1,18 @@
 export default {
   wrapper: 'relative',
-  base: 'relative block w-full disabled:cursor-not-allowed disabled:opacity-75 focus:outline-none border-0',
+  base: 'relative block w-full disabled:cursor-not-allowed disabled:opacity-75 focus:outline-none border-0 sm:text-sm sm:leading-6',
+  font: 'font-medium',
   form: 'form-input',
   rounded: 'rounded-md',
-  placeholder: 'placeholder-gray-400 dark:placeholder-gray-500',
+  placeholder: 'placeholder-gray-500 dark:placeholder-gray-600',
   size: {
-    '2xs': 'text-xs',
     xs: 'text-xs',
     sm: 'text-sm',
-    md: 'text-sm',
+    md: 'text-base',
     lg: 'text-sm',
     xl: 'text-base'
   },
   gap: {
-    '2xs': 'gap-x-1',
     xs: 'gap-x-1.5',
     sm: 'gap-x-1.5',
     md: 'gap-x-2',
@@ -21,16 +20,14 @@ export default {
     xl: 'gap-x-2.5'
   },
   padding: {
-    '2xs': 'px-2 py-1',
-    xs: 'px-2.5 py-1.5',
-    sm: 'px-2.5 py-1.5',
-    md: 'px-3 py-2',
-    lg: 'px-3.5 py-2.5',
+    xs: 'px-2 py-1',
+    sm: 'px-2 py-1',
+    md: 'px-2.5 py-1.5',
+    lg: 'px-3 py-2',
     xl: 'px-3.5 py-2.5'
   },
   leading: {
     padding: {
-      '2xs': 'ps-7',
       xs: 'ps-8',
       sm: 'ps-9',
       md: 'ps-10',
@@ -40,7 +37,6 @@ export default {
   },
   trailing: {
     padding: {
-      '2xs': 'pe-7',
       xs: 'pe-8',
       sm: 'pe-9',
       md: 'pe-10',
@@ -50,22 +46,21 @@ export default {
   },
   color: {
     white: {
-      outline: 'shadow-sm bg-white dark:bg-gray-900 text-gray-900 dark:text-white ring-1 ring-inset ring-gray-300 dark:ring-gray-700 focus:ring-2 focus:ring-primary-500 dark:focus:ring-primary-400'
+      outline: 'shadow-sm bg-white dark:bg-gray-900 text-gray-900 dark:text-white ring-1 ring-inset ring-gray-300 dark:ring-gray-700 focus:ring-2 focus:ring-primary-600 dark:focus:ring-primary-500'
     },
     gray: {
-      outline: 'shadow-sm bg-gray-50 dark:bg-gray-800 text-gray-900 dark:text-white ring-1 ring-inset ring-gray-300 dark:ring-gray-700 focus:ring-2 focus:ring-primary-500 dark:focus:ring-primary-400'
+      outline: 'shadow-sm bg-gray-50 dark:bg-gray-800 text-gray-900 dark:text-white ring-1 ring-inset ring-gray-300 dark:ring-gray-700 focus:ring-2 focus:ring-primary-600 dark:focus:ring-primary-500'
     }
   },
   variant: {
-    outline: 'shadow-sm bg-transparent text-gray-900 dark:text-white ring-1 ring-inset ring-{color}-500 dark:ring-{color}-400 focus:ring-2 focus:ring-{color}-500 dark:focus:ring-{color}-400',
+    outline: 'shadow-sm bg-transparent text-gray-900 dark:text-white ring-1 ring-inset ring-primary-600 dark:ring-primary-500 focus:ring-2 focus:ring-primary-600 dark:focus:ring-primary-500',
     none: 'bg-transparent focus:ring-0 focus:shadow-none'
   },
   icon: {
-    base: 'flex-shrink-0 text-gray-400 dark:text-gray-500',
-    color: 'text-{color}-500 dark:text-{color}-400',
+    base: 'flex-shrink-0 text-gray-500 dark:text-gray-600',
+    color: 'text-primary-600 dark:text-primary-500',
     loading: 'animate-spin',
     size: {
-      '2xs': 'h-4 w-4',
       xs: 'h-4 w-4',
       sm: 'h-5 w-5',
       md: 'h-5 w-5',
@@ -76,7 +71,6 @@ export default {
       wrapper: 'absolute inset-y-0 start-0 flex items-center',
       pointer: 'pointer-events-none',
       padding: {
-        '2xs': 'px-2',
         xs: 'px-2.5',
         sm: 'px-2.5',
         md: 'px-3',
@@ -88,7 +82,6 @@ export default {
       wrapper: 'absolute inset-y-0 end-0 flex items-center',
       pointer: 'pointer-events-none',
       padding: {
-        '2xs': 'px-2',
         xs: 'px-2.5',
         sm: 'px-2.5',
         md: 'px-3',
@@ -98,8 +91,8 @@ export default {
     }
   },
   default: {
-    size: 'sm',
-    color: 'white',
+    size: 'md',
+    color: 'gray',
     variant: 'outline',
     loadingIcon: 'i-heroicons-arrow-path-20-solid'
   }